# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /resources:
    get:
      operationId: listResources
      tags: [resources]
      summary: List published resources (+ own unpublished)
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
            description: Comma-separated tag list
        - name: type
          in: query
          schema:
            type: string
            enum: [link, file, course]
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Paginated resources
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResourceListResponse"
    post:
      operationId: createResource
      tags: [resources]
      summary: Submit a new resource (published=false by default)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateResourceBody"
      responses:
        "201":
          description: Created resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /resources/{slug}:
    get:
      operationId: getResource
      tags: [resources]
      summary: Get a single resource by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
    delete:
      operationId: deleteResource
      tags: [resources]
      summary: Delete a resource (author or admin)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /resources/{slug}/file:
    post:
      operationId: uploadResourceFile
      tags: [resources]
      summary: Upload file for a resource of type=file (multipart)
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "200":
          description: Updated resource with filePath set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
...
  /admin/resources/{slug}/publish:
    post:
      operationId: adminPublishResource
      tags: [resources]
      summary: Approve and publish a resource
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Resource"
  /admin/resources/{slug}/reject:
    post:
      operationId: adminRejectResource
      tags: [resources]
      summary: Reject and delete a resource
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminRejectResourceBody"
      responses:
        "204":
          description: Rejected and deleted

  # ── Marketplace ──────────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
...
  /notifications/{notificationId}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a single notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Marked

  # ── Admin ─────────────────────────────────────────────────────────────────────
...
  /storage/objects/{objectPath}:
    get:
      operationId: getStorageObject
      tags: [storage]
      summary: Serve a stored object
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Object file
components:
  schemas:
...
    Resource:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [link, file, course]
        url:
          type: string
          nullable: true
        filePath:
          type: string
          nullable: true
        description:
          type: string
        coverUrl:
          type: string
          nullable: true
        authorId:
          type: string
        authorName:
          type: string
        authorAvatar:
          type: string
          nullable: true
        published:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, type, description, authorId, authorName, published, tags, createdAt]
    ResourceListResponse:
      type: object
      properties:
        resources:
          type: array
          items:
            $ref: "#/components/schemas/Resource"
        nextCursor:
          type: string
          nullable: true
      required: [resources]
    CreateResourceBody:
      type: object
      properties:
        title:
          type: string
        type:
          type: string
          enum: [link, file, course]
        url:
          type: string
          nullable: true
        description:
          type: string
        coverUrl:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
      required: [title, type]
    AdminRejectResourceBody:
      type: object
      properties:
        reason:
          type: string
      required: [reason]

    # ── Marketplace ────────────────────────────────────────────────────────────
...
    Notification:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        type:
          type: string
          enum: [event_rsvp, forum_reply, forum_reaction, message_received, resource_liked, admin_action]
        title:
          type: string
        body:
          type: string
        link:
          type: string
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required: [id, userId, type, title, body, isRead, createdAt]
    NotificationListResponse:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        unreadCount:
          type: integer
      required: [notifications, unreadCount]

    # ── Admin ──────────────────────────────────────────────────────────────────
...